Renamed locale and currency data objects (no more use of "bag")
